Fixes TimSort bug
Fixes issue #2, addressing the bug explained in [this blog
post](http://envisage-project.eu/proving-android-java-and-python-sorting-algorithm-is-broken-and-how-to-fix-it/).

[According to Tim Peters](http://bugs.python.org/issue23515#msg236603),
this is a dumb way to fix the bug. [According to the original post
authors](http://envisage-project.eu/timsort-specification-and-verification/),
this is suboptimal but correct.

I've elected to use the suboptimal Java fix so that I can more easily track
the JDK code. I've also taken this opportunity to pull in the Oracle
comments in order to make the diff as small as possible.