Simplifies POM
In my former attempts with `maven-source-plugin\', I was assuming that it
would put the contents of `target/generated-sources` into the source jar.
What apparently happens in actuality is that `fmpp-maven-plugin` registers
that it output source code to its `outputDirectory`. Since the old
`outputDirectory` path included the package directories, the packaging was
broken.

By moving the template files to `fmpp/templates/net/mintern/primitive`, we
end up putting the generated source files into the correct hierarchy without
including the package hierarchy in the `outputDirectory` parameter. As a
result, the configuration is simplified and the build works correctly.